feature: 레이아웃 컴포넌트 완성 #28
