+++ src/components/Header.tsx
@@ -1,91 +1,19 @@
 /** @jsxImportSource @emotion/react */
-import { css, SerializedStyles } from '@emotion/react';
-import { useLocation } from 'react-router-dom';
-import { Home, Calendar, WalletMinimal, User } from 'lucide-react';
+import { css } from '@emotion/react';
 import { colors } from '@/constants/colors';
-import NavItem from '@/components/NavItem';
 
-const Header = ({ css }: { css?: SerializedStyles }) => {
-	const location = useLocation();
-
-	return (
-		<header css={[headerStyle, css]}>
-			<nav css={navStyle}>
-				<ul css={navListStyle}>
-					<NavItem
-						to="/"
-						icon={
-							<Home
-								css={iconStyle(
-									location.pathname === '/' ||
-										location.pathname.startsWith('/home'),
-								)}
-							/>
-						}
-						label="홈"
-						isActive={
-							location.pathname === '/' || location.pathname.startsWith('/home')
-						}
-					/>
-					<NavItem
-						to="/schedule"
-						icon={
-							<Calendar css={iconStyle(location.pathname.startsWith('/schedule'))} />
-						}
-						label="일정표"
-						isActive={location.pathname.startsWith('/schedule')}
-					/>
-					<NavItem
-						to="/wage"
-						icon={
-							<WalletMinimal css={iconStyle(location.pathname.startsWith('/wage'))} />
-						}
-						label="급여"
-						isActive={location.pathname.startsWith('/wage')}
-					/>
-					<NavItem
-						to="/profile"
-						icon={<User css={iconStyle(location.pathname.startsWith('/profile'))} />}
-						label="프로필"
-						isActive={location.pathname.startsWith('/profile')}
-					/>
-				</ul>
-			</nav>
-		</header>
-	);
-};
+const Header = () => (
+	<header css={headerStyle}>
+		<h2>우아한 치즈케잌</h2>
+	</header>
+);
 
 export default Header;
 
 const headerStyle = css`
-	position: fixed;
-	bottom: 0;
-	width: 100%;
-	max-width: 430px;
-	height: 60px;
-	background-color: ${colors.veryLightGray};
+	height: 52px;
+	background-color: ${colors.primaryYellow};
 	display: flex;
 	align-items: center;
 	justify-content: center;
-	border-top: 1px solid ${colors.lightGray};
-`;
-
-const navStyle = css`
-	display: flex;
-	padding: 6px 0 12px 0;
-`;
-
-const navListStyle = css`
-	display: flex;
-	gap: 54px;
-	list-style: none;
-	padding: 0;
-	margin: 0;
-`;
-
-const iconStyle = (isActive: boolean) => css`
-	width: 26px;
-	height: 26px;
-	stroke-width: 1.4;
-	color: ${isActive ? colors.black : colors.gray};
 `;
